requirements.txt
feedparser==6.0.10
requests==2.31.0
google-generativeai==0.8.3
python-dotenv==1.0.0 